+++ firebase-ai/app/src/main/java/com/google/firebase/quickstart/ai/feature/live/StreamRealtimeScreen.kt
@@ -37,8 +37,6 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.serialization.Serializable
 
-@Serializable
-class StreamRealtimeRoute(val sampleId: String)
 
 @RequiresPermission(Manifest.permission.RECORD_AUDIO)
 @Composable
