+++ firebase-ai/app/src/main/java/com/google/firebase/quickstart/ai/feature/text/ChatScreen.kt
@@ -79,44 +79,6 @@ import com.google.firebase.ai.type.WebGroundingChunk
 import kotlinx.coroutines.launch
 import kotlinx.serialization.Serializable
 
-@Serializable
-class ChatRoute(val sampleId: String? = null)
-
-@Serializable
-object TravelTipsRoute
-
-@Serializable
-object WeatherChatRoute
-
-@Serializable
-object CourseRecommendationsRoute
-
-@Serializable
-object AudioSummarizationRoute
-
-@Serializable
-object AudioTranslationRoute
-
-@Serializable
-object ImageBlogCreatorRoute
-
-@Serializable
-object ImageGenerationRoute
-
-@Serializable
-object DocumentComparisonRoute
-
-@Serializable
-object VideoHashtagGeneratorRoute
-
-@Serializable
-object VideoSummarizationRoute
-
-@Serializable
-object GoogleSearchGroundingRoute
-
-@Serializable
-object ThinkingChatRoute
 
 @Composable
 fun ChatScreen(
