+++ firebase-ai/app/src/main/java/com/google/firebase/quickstart/ai/feature/text/ImageGenerationViewModel.kt
@@ -1,11 +1,16 @@
 package com.google.firebase.quickstart.ai.feature.text
 
+import kotlinx.serialization.Serializable
+
 import com.google.firebase.Firebase
 import com.google.firebase.ai.Chat
 import com.google.firebase.ai.ai
 import com.google.firebase.ai.type.Content
 import com.google.firebase.ai.type.ResponseModality
 import com.google.firebase.ai.type.generationConfig
+
+@Serializable
+object ImageGenerationRoute
 
 class ImageGenerationViewModel : ChatViewModel() {
